requirements.txt
pyserial
slack_sdk